# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Keasy API",
    "description": "Keasy — data pipeline execution and cloud storage API",
    "contact": {
      "name": "Ángel Iglesias Préstamo",
      "email": "angel.iglesias.prestamo@gmail.com"
    },
    "license": {
      "name": ""
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/v1/assistant/generate-stream": {
      "post": {
        "tags": [
          "Assistant"
        ],
        "operationId": "generate_script_stream",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "SSE stream: delta events + complete with GenerateResponse"
          },
          "400": {
            "description": "AI provider not configured"
          }
        }
      }
    },
    "/v1/assistant/suggest-stream": {
      "post": {
        "tags": [
          "Assistant"
        ],
        "operationId": "suggest_cqs_stream",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SuggestRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "SSE stream: delta events + complete with SuggestResponse"
          },
          "400": {
            "description": "AI provider not configured"
          }
        }
      }
    },
...
    "/v1/cloud-accounts": {
      "get": {
        "tags": [
          "Cloud Accounts"
        ],
        "operationId": "list_accounts",
        "responses": {
          "200": {
            "description": "List of cloud accounts",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CloudAccountSummary"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Cloud Accounts"
        ],
        "operationId": "create_account",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCloudAccountRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Cloud account created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CloudAccountSummary"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed"
          }
        }
      }
    },
...
    "/v1/connections": {
      "get": {
        "tags": [
          "Connections"
        ],
        "operationId": "list_connections",
        "parameters": [
          {
            "name": "type",
            "in": "query",
            "description": "Filter by connection type",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of connections",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Connection"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Connections"
        ],
        "operationId": "create_connection",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateConnectionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Connection created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Connection"
                }
              }
            }
          },
          "400": {
            "description": "Invalid connection or container not found"
          }
        }
      }
    },
...
    "/v1/connections/urls": {
      "post": {
        "tags": [
          "Connections"
        ],
        "summary": "Sign GET URLs for locators, each with the credentials of the connection it\nlies under. Serves the editor and a job's run alike: a member reads what\nevery source connection holds, and the sink only through the job that\nwrote it. Public HTTP locators come back as they are.",
        "operationId": "sign_locators",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignLocatorsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Signed GET URLs for the locators the caller may read",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SignLocatorsResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/jobs": {
      "get": {
        "tags": [
          "Jobs"
        ],
        "operationId": "list_jobs",
        "responses": {
          "200": {
            "description": "The caller's jobs",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Job"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Jobs"
        ],
        "operationId": "create_job",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateJobRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Draft job created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Job"
                }
              }
            }
          },
          "202": {
            "description": "Job submitted for execution",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Job"
                }
              }
            }
          },
          "400": {
            "description": "The destination is not a sink"
          }
        }
      }
    },
    "/v1/jobs/{id}": {
      "get": {
        "tags": [
          "Jobs"
        ],
        "operationId": "get_job",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Job ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Job details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Job"
                }
              }
            }
          },
          "404": {
            "description": "Job not found"
          }
        }
      },
      "put": {
        "tags": [
          "Jobs"
        ],
        "operationId": "update_job",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Job ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateJobRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Job updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Job"
                }
              }
            }
          },
          "400": {
            "description": "Job is not a draft"
          },
          "404": {
            "description": "Job not found"
          }
        }
      },
      "delete": {
        "tags": [
          "Jobs"
        ],
        "operationId": "delete_job",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Job ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Job deleted"
          },
          "404": {
            "description": "Job not found"
          },
          "409": {
            "description": "Job is still running"
          }
        }
      },
      "patch": {
        "tags": [
          "Jobs"
        ],
        "summary": "The browser ran the mapping and uploaded the output; this records the\noutcome. `Completed` stores the run report verbatim, unread.",
        "operationId": "complete_job",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Job ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompleteJobRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Job status updated from the browser run",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Job"
                }
              }
            }
          },
          "404": {
            "description": "Job not found"
          }
        }
      }
    },
...
    "/v1/jobs/{id}/discover/urls": {
      "post": {
        "tags": [
          "Discovery"
        ],
        "summary": "Sign GET URLs so the browser reads the dataset directly — the reading twin of\n[`resolve_output_urls`]. It takes the list the corpus reader enumerated and\nderives none.",
        "operationId": "resolve_discover_urls",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Job ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DatasetUrlsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Signed GET URLs for the requested dataset keys",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResolveResponse"
                }
              }
            }
          },
          "400": {
            "description": "The destination connection is gone, or a path outside the dataset"
          },
          "404": {
            "description": "Job not found"
          }
        }
      }
    },
    "/v1/jobs/{id}/output/urls": {
      "post": {
        "tags": [
          "Discovery"
        ],
        "summary": "Sign PUT URLs so the browser uploads the output it just produced straight to\nthe job's sink.",
        "operationId": "resolve_output_urls",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Job ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DatasetUrlsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Signed PUT URLs for the output keys",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResolveResponse"
                }
              }
            }
          },
          "400": {
            "description": "The destination connection is gone, or a path outside the dataset"
          },
          "404": {
            "description": "Job not found"
          }
        }
      }
    },
    "/v1/jobs/{id}/relations": {
      "put": {
        "tags": [
          "Jobs"
        ],
        "summary": "What the corpus reader found: the relations a finished job's output holds,\ntheir names and the files that carry them, as `@fossil-lang/corpus`\nenumerated them in the browser.",
        "description": "It is a second call and not a field of the completion because naming a\nrelation is an answer only a reader holding the manifests can give, and the\nrun report is not that reader. keasy stores the answer and registers the\ndataset in the DuckLake catalog by reference — one atomic snapshot,\nidempotent, composing nothing: every name and every path in that SQL came\nfrom this payload.",
        "operationId": "publish_relations",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Job ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PublishRelationsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Relations stored and the dataset registered",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Job"
                }
              }
            }
          },
          "400": {
            "description": "A file path outside the dataset"
          },
          "404": {
            "description": "Job not found"
          }
        }
      }
    },
    "/v1/org/identity": {
      "get": {
        "tags": [
          "Organization"
        ],
        "operationId": "get_org_identity",
        "responses": {
          "200": {
            "description": "Workspace identity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrgIdentity"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Organization"
        ],
        "operationId": "update_org_identity",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrgIdentity"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Identity updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrgIdentity"
                }
              }
            }
          },
          "400": {
            "description": "Validation error"
          }
        }
      }
    },
...
    "/v1/settings/catalog-storage": {
      "get": {
        "tags": [
          "Settings"
        ],
        "operationId": "get_catalog_storage",
        "responses": {
          "200": {
            "description": "Catalog storage config",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogStoragePayload"
                }
              }
            }
          },
          "204": {
            "description": "Not configured"
          }
        }
      },
      "put": {
        "tags": [
          "Settings"
        ],
        "operationId": "save_catalog_storage",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CatalogStoragePayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Catalog storage saved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogStoragePayload"
                }
              }
            }
          },
          "400": {
            "description": "Validation error"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CloudAccountSummary": {
        "type": "object",
        "required": [
          "id",
          "name",
          "provider_id",
          "fields"
        ],
        "properties": {
          "auth_method": {
            "type": [
              "string",
              "null"
            ]
          },
          "fields": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            },
            "propertyNames": {
              "type": "string"
            }
          },
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "provider_id": {
            "type": "string"
          }
        }
      },
...
      "CreateCloudAccountRequest": {
        "type": "object",
        "description": "Every field value arrives as a secret: which of them are credentials is the\nprovider schema's to say, and until it has, none is logged or printed.",
        "required": [
          "name",
          "provider_id",
          "fields"
        ],
        "properties": {
          "auth_method": {
            "type": [
              "string",
              "null"
            ]
          },
          "fields": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            },
            "propertyNames": {
              "type": "string"
            }
          },
          "name": {
            "type": "string"
          },
          "provider_id": {
            "type": "string"
          }
        }
      },
...
      "GenerateRequest": {
        "type": "object",
        "required": [
          "domain",
          "competency_questions",
          "schemas"
        ],
        "properties": {
          "competency_questions": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "domain": {
            "type": "string"
          },
          "schemas": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FileSchema"
            }
          }
        }
      },
...
      "SuggestRequest": {
        "type": "object",
        "required": [
          "domain",
          "schemas"
        ],
        "properties": {
          "domain": {
            "type": "string"
          },
          "schemas": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FileSchema"
            }
          }
        }
      },
...
    }
  }
}